i18n: 多言語リソースの更新とi18n定義の拡充
