+++ crates/katana-ui/src/i18n/types/action.rs
@@ -1,0 +1,34 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ActionMessages {
+    pub expand_all: String,
+    pub collapse_all: String,
+    pub collapse_sidebar: String,
+    pub refresh_explorer: String,
+    pub toggle_filter: String,
+    pub remove_workspace: String,
+    pub recursive_expand: String,
+    pub recursive_open_all: String,
+    pub toggle_toc: String,
+    pub show_meta_info: String,
+    pub new_file: String,
+    pub new_directory: String,
+    pub open: String,
+    pub rename: String,
+    pub delete: String,
+    pub copy_path: String,
+    pub reveal_in_os: String,
+    pub save: String,
+    pub refresh_document: String,
+    pub cancel: String,
+    pub discard: String,
+    pub confirm: String,
+    pub copy_relative_path: String,
+    pub toggle_slideshow: String,
+    pub open_file: String,
+    pub save_file: String,
+    pub new_workspace: String,
+    pub open_workspace: String,
+    pub copy: String,
+}
